requirements.txt
arrow==1.2.2
asgiref==3.5.0
backports.zoneinfo==0.2.1
blessed==1.19.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-phonenumber-field==6.1.0
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==4.0.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mccabe==0.6.1
packaging==21.3
phonenumbers==8.12.43
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
redis==3.5.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
